exp(x) Transformation

Perform a exp(x) transformation

exp_x(x, standardize = TRUE, warn = TRUE, ...)

# S3 method for exp_x
predict(object, newdata = NULL, inverse = FALSE, ...)

# S3 method for exp_x
print(x, ...)

Arguments

x

A vector to normalize with with x

standardize

If TRUE, the transformed values are also centered and scaled, such that the transformation attempts a standard normal

warn

Should a warning result from infinite values?

...

additional arguments

object

an object of class 'exp_x'

newdata

a vector of data to be (potentially reverse) transformed

inverse

if TRUE, performs reverse transformation

Value

A list of class exp_x with elements

x.t

transformed original data

x

original data

mean

mean after transformation but prior to standardization

sd

sd after transformation but prior to standardization

n

number of nonmissing observations

norm_stat

Pearson's P / degrees of freedom

standardize

was the transformation standardized

The predict function returns the numeric value of the transformation performed on new data, and allows for the inverse transformation as well.

Details

exp_x performs a simple exponential transformation in the context of bestNormalize, such that it creates a transformation that can be estimated and applied to new data via the predict function.

Examples

x <- rgamma(100, 1, 1)

exp_x_obj <- exp_x(x)
exp_x_obj
#> Standardized exp(x) Transformation with 100 nonmissing obs.:
#>  Relevant statistics:
#>  - mean (before standardization) = 8.930029 
#>  - sd (before standardization) = 40.6462 
p <- predict(exp_x_obj)
x2 <- predict(exp_x_obj, newdata = p, inverse = TRUE)

all.equal(x2, x)
#> [1] TRUE